+++ test/libsolidity/semanticTests/expressions/uncalled_address_transfer_send.sol
@@ -1,7 +1,7 @@
 contract TransferTest {
 	fallback() external payable {
 		// This used to cause an ICE
-		address(this).transfer;
+		payable(this).transfer;
 	}
 
 	function f() pure public {}