+++ test/libsolidity/semanticTests/functionCall/send_zero_ether.sol
@@ -10,7 +10,7 @@ contract Main {
 
     function s() public returns (bool) {
         Receiver r = new Receiver();
-        return address(r).send(0);
+        return payable(r).send(0);
     }
 }