+++ test/libsolidity/syntaxTests/types/address/address_constant.sol
@@ -1,6 +1,6 @@
 contract C {
     address constant a = address(0);
-    address payable constant b = address(0);
+    address payable constant b = payable(0);
     function f() public pure returns (address, address) {
         return (a,b);
     }