+++ test/libsolidity/syntaxTests/types/address/address_to_contract_payable_fallback.sol
@@ -1,6 +1,6 @@
 contract C {
   function f() public pure returns (C c) {
-    c = C(address(2));
+    c = C(payable(2));
   }
   fallback() external payable {
   }