+++ test/libsolidity/syntaxTests/types/address/address_tuple_fail.sol
@@ -1,6 +1,6 @@
 contract C {
     function f() public view returns (address payable a, address b) {
-        (address c, address payable d) = (address(this), address(0));
+        (address c, address payable d) = (address(this), payable(0));
         (a,b) = (c,d);
     }
 }