+++ test/libsolidity/syntaxTests/types/address/literal_to_address_call.sol
@@ -1,0 +1,5 @@
+contract C {
+    function f() public returns (bool success) {
+        (success, ) = (address(0)).call{value: 30}("");
+    }
+}